Rename blocks.js to views.js
